neatogen: represent 'nsites' as a 'size_t'
This squashes 3 -Wsign-conversion warnings. Note that this required also
converting `nvertices` and `Site.sitenbr` to `size_t` too.

+++ lib/neatogen/geometry.c
@@ -10,7 +10,7 @@
 
 #include <neatogen/geometry.h>
 #include <math.h>
-
+#include <stddef.h>
 
 Point origin = { 0, 0 };
 
@@ -18,7 +18,7 @@ double xmin, xmax, ymin, ymax;	/* min and max x and y values of sites */
 double deltax,			/* xmax - xmin */
  deltay;			/* ymax - ymin */
 
-int nsites;
+size_t nsites;
 int sqrt_nsites;
 
 void geominit()
